Update Charter.md
Updated the Mission statement (1.) per consortium member feedback to soften its claims.  The Mission now reads as follows:
The mission of the OCSF is to create an industry-wide framework for data exchange, aggregation and analysis, principally but not exclusively for cybersecurity use cases. The OCSF is centered on its strongly typed data structures, standard attribute dictionary and taxonomy structure, and extensible core schema.